refactor(client): show the purchase nudge on the home page, not in-game
The spawn phase is the most disruptive moment to interrupt a player.
Move the nudge out of the HUD: it is now a home-page component that
reacts to the userMeResponse event on page load (the same pattern as the
marketing consent toast), with the same eligibility gates. "Visit store"
opens the store modal in place instead of a new tab.
